+++ src/components/vector-properties/VcsVectorPropertiesComponent.vue
@@ -462,7 +462,6 @@
     </v-container>
   </VcsFormSection>
 </template>
-
 <script>
   import { computed } from 'vue';
   import { VContainer, VRow, VCol, VDivider } from 'vuetify/lib';